refactor: Change project structure
