+++ kotlin-spark-api/2.4/src/test/kotlin/org/jetbrains/kotlinx/spark/api/TypeInferenceTest.kt
